feat: add lh fir splitter

+++ source/dsp/delay/integer_delay.hpp
@@ -92,7 +92,7 @@ namespace zldsp::delay {
         }
 
         void setDelayInSamples(const int delay_samples) {
-            setDelay(static_cast<FloatType>(delay_samples) / static_cast<FloatType>(sample_rate_));
+            setDelay(static_cast<FloatType>(static_cast<double>(delay_samples) / sample_rate_));
         }
 
         [[nodiscard]] int getDelayInSamples() const {
